feat(stage81): 实现 AI 增强平台核心模块
- 新增 ai/smart_debugger.rs: AI 智能调试器
  * 错误诊断和根因分析
  * 修复建议生成
  * 断点建议和调试路径优化

- 新增 ai/auto_optimizer.rs: 自动性能优化器
  * 实时性能分析和热点检测
  * 性能瓶颈识别
  * 自动优化建议和内存优化

- 新增 ai/predictive_scaler.rs: 预测性扩展器
  * 资源使用预测和趋势分析
  * 自动扩展策略和智能调度
  * 异常检测和性能预测

- 更新 ai/mod.rs 导出所有新模块
- 修复 llm_engine.rs 和 model_manager.rs 中的 Runtime 类型引用

实现 Stage 81 AI 增强平台的核心功能，为开发者提供智能化的代码生成、调试和优化能力。

+++ src/ecosystem/community/portal.rs
@@ -68,7 +68,7 @@ pub struct ModuleInfo {
 }
 
 /// 模块类别
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum ModuleCategory {
     WebFramework,
     Utility,
@@ -299,12 +299,12 @@ impl CommunityPortal {
 }
 
 /// 模块评分摘要
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default)]
 pub struct ModuleRatingSummary {
     pub module_id: String,
     pub total_ratings: usize,
     pub average_rating: f64,
-    pub rating_distribution: HashMap<usize, usize>,
+    pub rating_distribution: HashMap<u8, usize>,
 }
 
 /// 模块搜索结果
